feat: Reorganize folder structure #8
