+++ src/app/[locale]/search/page.tsx
