+++ src/lib/i18n/navigation.ts
@@ -1,0 +1,6 @@
+/* eslint-disable no-unused-expressions */
+import { createSharedPathnamesNavigation } from "next-intl/navigation";
+import { locales, localePrefix } from "@/lib/i18n/config";
+
+export const { Link, redirect, usePathname, useRouter } =
+  createSharedPathnamesNavigation({ locales, localePrefix });
