--- src/navigation.ts
@@ -1,10 +1,0 @@
-/* eslint-disable no-unused-expressions */
-import { createLocalizedPathnamesNavigation } from "next-intl/navigation";
-import { locales, pathnames, localePrefix } from "@/lib/i18n/config";
-
-export const { Link, redirect, usePathname, useRouter } =
-  createLocalizedPathnamesNavigation({
-    locales,
-    pathnames,
-    localePrefix,
-  });
